add delete all grups from all users and add Not Found Exception to Groups

+++ src/main/java/com/usergroup/api/users/service/UserService.java
@@ -41,8 +41,14 @@ public class UserService {
         }
     }
 
-    public void delete(Long id) {
-        userRepository.deleteById(id);
+    public Boolean delete(Long id) {
+        if  (getUserById(id)!= null) {
+            userRepository.deleteById(id);
+            return true;
+        } else{
+            return false;
+        }
+
     }
